# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Presensi Event Alumni Pesantren API",
        "description": "API documentation for Presensi Event Alumni Pesantren system.",
        "contact": {
            "email": "admin@pesantren.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/admin/events/{id}/qr": {
            "get": {
                "tags": [
                    "Admin - Event QR Management"
                ],
                "summary": "Get active event QR code",
                "description": "Returns active QR code data for a specific event. Admin only.",
                "operationId": "adminGetEventQrCode",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Active QR code data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "qr_code": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "event_id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "qr_token": {
                                                            "type": "string",
                                                            "example": "550e8400-e29b-41d4-a716-446655440000"
                                                        },
                                                        "qr_code_image": {
                                                            "type": "string",
                                                            "example": "qrcodes/550e8400.svg"
                                                        },
                                                        "qr_code_url": {
                                                            "type": "string",
                                                            "example": "http://localhost:8000/storage/qrcodes/550e8400.svg"
                                                        },
                                                        "timeout_minutes": {
                                                            "type": "integer",
                                                            "example": 60
                                                        },
                                                        "is_active": {
                                                            "type": "boolean",
                                                            "example": true
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "example": "2026-05-25T13:56:00.000000Z"
                                                        },
                                                        "expired_at": {
                                                            "type": "string",
                                                            "example": "2026-05-25T14:56:00.000000Z"
                                                        },
                                                        "is_expired": {
                                                            "type": "boolean",
                                                            "example": false
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event or QR code not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/events/{id}/qr/generate": {
            "post": {
                "tags": [
                    "Admin - Event QR Management"
                ],
                "summary": "Generate event QR code",
                "description": "Generates a new QR code for a specific event. Previous active QR code will be deactivated. Admin only.",
                "operationId": "adminGenerateEventQrCode",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "timeout_minutes"
                                ],
                                "properties": {
                                    "timeout_minutes": {
                                        "type": "integer",
                                        "example": 60
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "QR code generated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "QR code generated successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "qr_code": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "event_id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "qr_token": {
                                                            "type": "string",
                                                            "example": "550e8400-e29b-41d4-a716-446655440000"
                                                        },
                                                        "qr_code_image": {
                                                            "type": "string",
                                                            "example": "qrcodes/550e8400.svg"
                                                        },
                                                        "qr_code_url": {
                                                            "type": "string",
                                                            "example": "http://localhost:8000/storage/qrcodes/550e8400.svg"
                                                        },
                                                        "valid_from": {
                                                            "type": "string",
                                                            "example": "2026-05-25T13:56:00.000000Z"
                                                        },
                                                        "timeout_minutes": {
                                                            "type": "integer",
                                                            "example": 60
                                                        },
                                                        "is_active": {
                                                            "type": "boolean",
                                                            "example": true
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "example": "2026-05-25T13:56:00.000000Z"
                                                        },
                                                        "expired_at": {
                                                            "type": "string",
                                                            "example": "2026-05-25T14:56:00.000000Z"
                                                        },
                                                        "is_valid_now": {
                                                            "type": "boolean",
                                                            "example": true
                                                        },
                                                        "is_expired": {
                                                            "type": "boolean",
                                                            "example": false
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/events/{id}/qr-image": {
            "get": {
                "tags": [
                    "Admin - Event QR Management"
                ],
                "summary": "Display active event QR Code image",
                "description": "Returns the active QR Code SVG image of a specific event directly in browser. Admin only.",
                "operationId": "adminGetGeneratedEventQrImage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "QR Code SVG image",
                        "content": {
                            "image/svg+xml": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event or QR code image not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error message here"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Validation failed"
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "The email field is required."
                            ]
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}